Reorganize removeUnused tests
- and add some documentation to explain that this rule can't work
yet on scala 3
