+++ scalafix-tests/input/src/main/scala-2.12/test/removeUnused/RemoveUnusedImportsTrailingCommas.scala
