+++ scalafix-tests/input/src/main/scala-2.13/test/removeUnused/RemoveUnusedParams.scala
@@ -1,7 +1,7 @@
 /*
 rule = RemoveUnused
  */
-package test
+package test.removeUnused
 
 object UnusedParams {
   def g(x: String => Unit): Unit = ???
