+++ scalafix-tests/input/src/main/scala-2.13/test/removeUnused/RemoveUnusedPatternVars.scala
@@ -1,7 +1,7 @@
 /*
 rule = RemoveUnused
  */
-package test
+package test.removeUnused
 
 case class AB(a: Int, b: String)
