+++ scalafix-tests/input/src/main/scala-2/test/removeUnused/RemoveUnusedImports2.scala
@@ -1,7 +1,7 @@
 /*
-rule = RemoveUnused
+rules = RemoveUnused
  */
-package test
+package test.removeUnused
 
 import scala.sys.process.{
   FileProcessLogger,
