+++ scalafix-tests/input/src/main/scala-2/test/removeUnused/RemoveUnusedImportsRename.scala
@@ -1,7 +1,7 @@
 /*
-rule = RemoveUnused
+rules = RemoveUnused
  */
-package test
+package test.removeUnused
 
 package t1 {
   import scala.collection.immutable.{List => IList}
