+++ scalafix-tests/input/src/main/scala-2/test/removeUnused/RemoveUnusedStatements.scala
@@ -1,6 +1,7 @@
 /*
-rule = RemoveUnused
+rules = RemoveUnused
 */
+package test.removeUnused
 
 class A {
   private val a = 1
