+++ scalafix-tests/input/src/main/scala-2/test/removeUnused/RemoveUnusedTerms.scala
@@ -1,7 +1,7 @@
 /*
-rule = RemoveUnused
+rules = RemoveUnused
  */
-package test
+package test.removeUnused
 
 object RemoveUnusedTerms {
