+++ scalafix-tests/output/src/main/scala-2.12/test/removeUnused/RemoveUnusedPatternVars.scala
