+++ scalafix-tests/output/src/main/scala-2.13/test/removeUnused/RemoveUnusedParams.scala
@@ -1,4 +1,4 @@
-package test
+package test.removeUnused
 
 object UnusedParams {
   def g(x: String => Unit): Unit = ???
