+++ scalafix-tests/output/src/main/scala-2.13/test/removeUnused/RemoveUnusedPatternVars.scala
@@ -1,4 +1,4 @@
-package test
+package test.removeUnused
 
 case class AB(a: Int, b: String)
