+++ scalafix-tests/output/src/main/scala-2/test/removeUnused/RemoveUnusedImports.scala
@@ -1,4 +1,4 @@
-package test
+package test.removeUnused
 
 import scala.util.control.NonFatal
 import scala.concurrent.Future
