+++ scalafix-tests/output/src/main/scala-2/test/removeUnused/RemoveUnusedImports2.scala
@@ -1,4 +1,4 @@
-package test
+package test.removeUnused
 
 import scala.sys.process.FileProcessLogger
 import scala.math.{
