+++ scalafix-tests/output/src/main/scala-2/test/removeUnused/RemoveUnusedStatements.scala
@@ -1,3 +1,5 @@
+package test.removeUnused
+
 class A {
