+++ scalafix-tests/output/src/main/scala-2/test/removeUnused/RemoveUnusedTerms.scala
@@ -1,4 +1,4 @@
-package test
+package test.removeUnused
 
 object RemoveUnusedTerms {
